fixed Issue 81: ChainReader CRD bug (Timestep problem)
ChainReader trajectory built from CRD contained garbage coordinates. This
was really due to the fact that when building a Timestep from a coordinate
array we did not enforce float32 type. This was fixed for base.Timestep and
and xdrfile.base.Timestep  and a test case was added.

git-svn-id: https://mdanalysis.googlecode.com/svn/trunk@914 dcee7792-c13e-0410-88f5-272e5f544b36

+++ MDAnalysis/coordinates/base.py
@@ -109,7 +109,7 @@ class Timestep(object):
             if arg.shape[1] == 3: self.numatoms = arg.shape[0]
             else: self.numatoms = arg.shape[0]   # Or should an exception be raised if coordinate structure is not 3-dimensional? Maybe velocities could be read one day... [DP]
 
-            self._pos = arg.copy('Fortran')
+            self._pos = arg.astype(numpy.float32).copy('Fortran',)
         else: raise ValueError("Cannot create an empty Timestep")
         self._x = self._pos[:,0]
         self._y = self._pos[:,1]
